pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
    "razorpay>=2.0.0",
    "werkzeug>=3.1.3",
]
